chore: signup query를 remotes 폴더 하위로 분리
