+++ src/features/auth/signup/components/SignupEmail.tsx
@@ -2,11 +2,11 @@ import SignForm from '../../SignForm';
 import envelopeWhiteIcon from '../../../../assets/icons/envelope-white.svg';
 import { css } from '@emotion/react';
 import { SubmitHandler, useFormContext } from 'react-hook-form';
-import usePostEmailVerification from '../remotes/hooks/usePostEmailVerification';
 import { useEffect } from 'react';
 import { SignupFields } from '../../../../pages/SignupPage';
 import { Spacing } from '../../../../shared/Spacing';
 import InputField from '../../../../shared/InputField';
+import usePostEmailVerification from '../remotes/postEmailVerification';
 
 interface SignupEmailProps {
   nextStep: () => void;
