--- src/features/auth/signup/remotes/hooks/usePostEmailVerification.ts
@@ -1,11 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-import { SignupRequest, postEmailVerification } from '../query';
-
-function usePostEmailVerification() {
-  return useMutation({
-    mutationKey: ['postEmailVerification'],
-    mutationFn: (data: Pick<SignupRequest, 'email'>) => postEmailVerification(data),
-  });
-}
-
-export default usePostEmailVerification;
