--- src/features/auth/signup/remotes/hooks/useSignup.ts
@@ -1,11 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-import { SignupRequest, postSignup } from '../query';
-
-function useSignup() {
-  return useMutation({
-    mutationKey: ['postSignup'],
-    mutationFn: (data: SignupRequest) => postSignup(data),
-  });
-}
-
-export default useSignup;
